my_first_api/requirements.txt
asgiref==3.8.1
Django==4.2.11
django-filter==24.2
django-guardian==2.4.0
djangorestframework==3.15.1
importlib-metadata==7.1.0
Markdown==3.6
pygments==2.17.2
PyYAML==6.0.1
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
zipp==3.18.1
